Add indication in UI when placements fail
If SeatStat cannot find a good seating chart for the
restrictions, the students for whom placements could
not be found will be colored red. This is a temporary
solution for the problem, and isn't very clear to users.
But I doubt users will encounter this in normal usage.
I need to do some more testing.